+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.10.0
 - feat: Retry fetching block after timeout. [PR 130](https://github.com/dariusc93/rust-ipfs/pull/130)
+- refactor: Use `Bytes` apart of unixfs operations. [PR 131](https://github.com/dariusc93/rust-ipfs/pull/131)
 - refactor: Remove option and use configuration directly. [PR 129](https://github.com/dariusc93/rust-ipfs/pull/129)
 - refactor: Remove option around chunker and use conversion for option arguments.
 - chore: Switch from libp2p-nat to libp2p-upnp. [PR 128](https://github.com/dariusc93/rust-ipfs/pull/128)